chore: 🔧 refactored functions from calendar utils as static methods of day class
